Always include `builtInMiddleware` in custom middleware types
